eo/requirements.txt
asgiref==3.8.1
Django>=4.2,<5.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Pillow==11.3.0
PyJWT==2.8.0
sqlparse==0.5.1